backend/src/main/kotlin/com/photograph/backend/config/security/component/Oauth2LogoutHandler.kt
package com.photograph.backend.config.security.component

import jakarta.servlet.http.HttpServletRequest
import jakarta.servlet.http.HttpServletResponse
import org.springframework.security.core.Authentication
import org.springframework.security.web.authentication.logout.CookieClearingLogoutHandler
import org.springframework.security.web.authentication.logout.LogoutSuccessHandler
import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler
import org.springframework.stereotype.Component

@Component
class Oauth2LogoutHandler : LogoutSuccessHandler {
    override fun onLogoutSuccess(
        request: HttpServletRequest,
        response: HttpServletResponse,
        authentication: Authentication
    ) {
        CookieClearingLogoutHandler("JSESSIONID").logout(request, response, authentication)
        SecurityContextLogoutHandler().logout(request, response, authentication)
        response.status = HttpServletResponse.SC_OK
    }
}
